Moved to a multi module model, added cert validation code
* Added code to verify client certificates based on ssh public keys
* Added an example client in examples module
* Renamed package from com.spotify.sshtlsclient to
  com.spotify.sshtls.{server,client,example}
