+++ pom.xml
@@ -3,10 +3,10 @@
   <modelVersion>4.0.0</modelVersion>
 
   <groupId>com.spotify</groupId>
-  <artifactId>ssh-tls-client</artifactId>
+  <artifactId>ssh-tls</artifactId>
   <version>0.1.0-SNAPSHOT</version>
-  <packaging>jar</packaging>
-  <name>ssh-tls-client</name>
+  <packaging>pom</packaging>
+  <name>ssh-tls</name>
   <description>A Java library that does a TLS handshake by using SSH keys.</description>
   <url>https://github.com/spotify/ssh-tls-client</url>
 
@@ -38,6 +38,13 @@
     </snapshotRepository>
   </distributionManagement>
 
+  <modules>
+    <module>client</module>
+    <module>validator</module>
+    <module>example</module>
+  </modules>
+
+
   <developers>
     <developer>
       <id>davidxia</id>
@@ -52,43 +59,17 @@
       <artifactId>slf4j-api</artifactId>
       <version>1.7.7</version>
     </dependency>
+
     <dependency>
       <groupId>com.google.guava</groupId>
       <artifactId>guava</artifactId>
       <version>17.0</version>
     </dependency>
-    <dependency>
-      <groupId>com.spotify</groupId>
-      <artifactId>ssh-agent-proxy</artifactId>
-      <version>0.1.2</version>
-    </dependency>
-    <dependency>
-      <groupId>org.bouncycastle</groupId>
-      <artifactId>bcprov-jdk15on</artifactId>
-      <version>1.53</version>
-    </dependency>
-    <dependency>
-      <groupId>org.bouncycastle</groupId>
-      <artifactId>bcpkix-jdk15on</artifactId>
-      <version>1.53</version>
-    </dependency>
-    <dependency>
-      <groupId>com.eaio.uuid</groupId>
-      <artifactId>uuid</artifactId>
-      <version>3.2</version>
-    </dependency>
 
-    <!-- test deps -->
     <dependency>
       <groupId>junit</groupId>
       <artifactId>junit</artifactId>
       <version>4.12</version>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.mockito</groupId>
-      <artifactId>mockito-all</artifactId>
-      <version>1.10.19</version>
       <scope>test</scope>
     </dependency>
   </dependencies>
