+++ validator/src/main/java/com/spotify/sshtls/validator/KeyProvider.java
@@ -1,0 +1,27 @@
+/*
+ * Copyright (c) 2015 Spotify AB.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.spotify.sshtls.validator;
+
+import java.security.PublicKey;
+
+
+/**
+ * Implementations of this class will provide an PublicKey instance given a username.
+ */
+public interface KeyProvider {
+  PublicKey getKey(String username);
+}
