+++ validator/src/test/resources/test_alice.crt
@@ -1,0 +1,20 @@
+-----BEGIN CERTIFICATE-----
+MIIDLTCCAhWgAwIBAgIIVszn33xI7hswDQYJKoZIhvcNAQEFBQAwNzELMAkGA1UE
+BhMCVVMxEDAOBgNVBAoMB1Nwb3RpZnkxFjAUBgNVBAMMDWhlbGlvcy1jbGllbnQw
+HhcNMTUxMTA1MTEyMDU0WhcNMTUxMTA3MTIyMDU0WjAXMRUwEwYKCZImiZPyLGQB
+AQwFYWxpY2UwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQDtsi2KpukV
+zMnOmGyT3pDwmyiVSIDmhj2J9t3b2XbASMy32Jm8ZrHnjKOsVQVBWaPDwlSvqPSC
+ME5cTMu/bLkQ6PZl2lsyvfmUQFmR2sFAy9M0SNjbllix2cxuxW3iB6m5I+67Hndx
+If0YCRBxkh3CKrUnqlJLkS8G6BhwSTufw6VKHsHQoMKA8sx0K3CuSpSC4UptBnjO
+zVfVe+Su/CxzGkidaZnSKs3Tqw48L4NYYczMZhhSZDOktgti/1wK43jWan8HjQ+w
+Xx15FVEKsXTMPncAZCjIw2OeZ42NinkWd+kk1ICOhJmAqLNLCJQp+GY7m5hJO2U+
+OQtXTUpeuOPVAgMBAAGjXTBbMB0GA1UdDgQWBBQyG/vCYNzaQ2vCJd8TjabrjQUk
+/TAfBgNVHSMEGDAWgBQyG/vCYNzaQ2vCJd8TjabrjQUk/TALBgNVHQ8EBAMCAoQw
+DAYDVR0TAQH/BAIwADANBgkqhkiG9w0BAQUFAAOCAQEAa82rgCpo/ZlrKywnQZdJ
+Fct+hcRXQyb5MC+/83XBVrdn1e0ay5Pp0ydbsSWqswv2pn7XqZB94/BCE45BHO3p
+7VcbaIwUYfXJvnixDlLUfUN0yl6GaUUd75uP5z25E1NN4Jquk9CbGHeDVFFdPXsg
+7cMX3uPTqRt5F0xRFuRU5yZxYiWA3wMFlSv+BSNyaZrcVOW1gwsbIkUW2cVGW4Mj
+ckkceDUnPOu9ynYvj1IipMWQq1Uy3Yk3Rm2D0tRGrtPecUMjHMIQej2+UB3aMlBP
+m087FpITdw3jiRABH0soX39aDH9hIaigxkA1Mchqwz3mitGd8Wa1NbAZhGmvcHnJ
+ag==
+-----END CERTIFICATE-----
