+++ validator/src/test/resources/test_noa.crt
@@ -1,0 +1,20 @@
+-----BEGIN CERTIFICATE-----
+MIIDKzCCAhOgAwIBAgIIdboh4IItEeUwDQYJKoZIhvcNAQEFBQAwNzELMAkGA1UE
+BhMCVVMxEDAOBgNVBAoMB1Nwb3RpZnkxFjAUBgNVBAMMDWhlbGlvcy1jbGllbnQw
+HhcNMTUxMTAzMTIxOTA2WhcNMTUxMTA1MTMxOTA2WjAVMRMwEQYKCZImiZPyLGQB
+AQwDbm9hMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA7bItiqbpFczJ
+zphsk96Q8JsolUiA5oY9ifbd29l2wEjMt9iZvGax54yjrFUFQVmjw8JUr6j0gjBO
+XEzLv2y5EOj2ZdpbMr35lEBZkdrBQMvTNEjY25ZYsdnMbsVt4gepuSPuux53cSH9
+GAkQcZIdwiq1J6pSS5EvBugYcEk7n8OlSh7B0KDCgPLMdCtwrkqUguFKbQZ4zs1X
+1XvkrvwscxpInWmZ0irN06sOPC+DWGHMzGYYUmQzpLYLYv9cCuN41mp/B40PsF8d
+eRVRCrF0zD53AGQoyMNjnmeNjYp5FnfpJNSAjoSZgKizSwiUKfhmO5uYSTtlPjkL
+V01KXrjj1QIDAQABo10wWzAdBgNVHQ4EFgQUMhv7wmDc2kNrwiXfE42m640FJP0w
+HwYDVR0jBBgwFoAUMhv7wmDc2kNrwiXfE42m640FJP0wCwYDVR0PBAQDAgKEMAwG
+A1UdEwEB/wQCMAAwDQYJKoZIhvcNAQEFBQADggEBAI3fCzF0pL1qBYBCxdi4AjIk
+mV1q5q/jW19pUMHBN6NbFIvNRum74dcf1UqPl8ek6qNlrMb+eNCO+DjUmr4b03A+
+tKRMUhSzmagxXmgO7bJWv8+EvfGK8JpWUWmYknbJNZonatoORomXkUSZjpqljdOk
+ryhaOMKlzxpg5iwlGtpjGgImd6O5D5+8YTpYiaKGxJH0XEn9LKUB1tnkfLduExYU
+uZ9ojaweKk5feMRBt8P/BtWFiRtFMEoir/RdlcA8tlxrV4rfT3GLbKOz0gdca6ib
+NZpIdC2krEKs33w54sixHFASNOkru1Yk7ImmD4eNiMsTZVOqBUVEAD4EttAMjfw=
+-----END CERTIFICATE-----
+
